refactor(formatter_css,formatter_graphql,formatter_yaml): move Gap utils to core (#25137)
CSS, GraphQL, YAML formatters had the same logic to inspect source blank lines.
